feat: Add border and elevation to TargetDialog component
This commit adds a border and elevation to the TargetDialog component in the TargetDialog.vue file. The border is set to a solid 2px green color (#00ff00) and the elevation is set to 24. This improves the visual appearance of the dialog and makes it stand out more on the page.